r1cs-mpc: mpc-prover: Implement transcript sharing
To construct valid Fiat-Shamir challenges without complex circuit
logic (i.e. implementing SHAKE over the authenticated field) we
instead open each transcript element before adding it to the
transcript. This way, all parties have a consistent view of the
transcript and the Fiat-Shamir challenges are public and accessible
to the verifier.